refactor: 유저가 소유한 공간에 예약이 있는지 확인하는 쿼리를 Member -> Reservation Repo로 이동
