+++ backend/src/test/java/com/woowacourse/zzimkkong/repository/MemberRepositoryTest.java
@@ -69,55 +69,4 @@ class MemberRepositoryTest extends RepositoryTest {
         assertThatThrownBy(() -> members.save(sameEmailMember))
                 .isInstanceOf(DataAccessException.class);
     }
-
-    @ParameterizedTest
-    @CsvSource({"true", "false"})
-    @DisplayName("멤버의 Id를 이용해 예약이 존재하는지 확인할 수 있다.")
-    void existsReservationsByMemberId(boolean isReservationExists) {
-        // given
-        Member savedMember = members.save(pobi);
-
-        Map luther = new Map(LUTHER_NAME, MAP_DRAWING_DATA, MAP_IMAGE_URL, pobi);
-        maps.save(luther);
-
-        Setting beSetting = Setting.builder()
-                .availableStartTime(BE_AVAILABLE_START_TIME)
-                .availableEndTime(BE_AVAILABLE_END_TIME)
-                .reservationTimeUnit(BE_RESERVATION_TIME_UNIT)
-                .reservationMinimumTimeUnit(BE_RESERVATION_MINIMUM_TIME_UNIT)
-                .reservationMaximumTimeUnit(BE_RESERVATION_MAXIMUM_TIME_UNIT)
-                .reservationEnable(BE_RESERVATION_ENABLE)
-                .enabledDayOfWeek(BE_ENABLED_DAY_OF_WEEK)
-                .build();
-
-        Space be = Space.builder()
-                .name(BE_NAME)
-                .color(BE_COLOR)
-                .description(BE_DESCRIPTION)
-                .area(SPACE_DRAWING)
-                .setting(beSetting)
-                .map(luther)
-                .build();
-
-        spaces.save(be);
-
-        if (isReservationExists) {
-            Reservation beAmZeroOne = Reservation.builder()
-                    .startTime(BE_AM_TEN_ELEVEN_START_TIME)
-                    .endTime(BE_AM_TEN_ELEVEN_END_TIME)
-                    .description(BE_AM_TEN_ELEVEN_DESCRIPTION)
-                    .userName(BE_AM_TEN_ELEVEN_USERNAME)
-                    .password(BE_AM_TEN_ELEVEN_PW)
-                    .space(be)
-                    .build();
-
-            reservations.save(beAmZeroOne);
-        }
-
-        // when
-        Boolean hasAnyReservations = members.existsReservationsByMember(savedMember);
-
-        // then
-        assertThat(hasAnyReservations).isEqualTo(isReservationExists);
-    }
 }
