+++ backend/src/test/java/com/woowacourse/zzimkkong/repository/ReservationRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.woowacourse.zzimkkong.repository;
 
 import com.woowacourse.zzimkkong.domain.*;
+import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -23,9 +24,11 @@ class ReservationRepositoryTest extends RepositoryTest {
     private Reservation beNextDayAmSixTwelve;
     private Reservation fe1ZeroOne;
 
+    private Member pobi;
+
     @BeforeEach
     void setUp() {
-        Member pobi = new Member(EMAIL, PW, ORGANIZATION);
+        pobi = new Member(EMAIL, PW, ORGANIZATION);
         Map luther = new Map(LUTHER_NAME, MAP_DRAWING_DATA, MAP_IMAGE_URL, pobi);
 
         Setting beSetting = Setting.builder()
@@ -223,5 +226,57 @@ class ReservationRepositoryTest extends RepositoryTest {
                 minimumDateTime,
                 maximumDateTime
         );
+    }
+
+    @ParameterizedTest
+    @CsvSource({"true", "false"})
+    @DisplayName("멤버의 Id를 이용해 예약이 존재하는지 확인할 수 있다.")
+    void existsReservationsByMemberId(boolean isReservationExists) {
+        // given
+        Member sakjung = new Member(NEW_EMAIL, PW, ORGANIZATION);
+        Member savedMember = members.save(sakjung);
+
+        Map luther = new Map(LUTHER_NAME, MAP_DRAWING_DATA, MAP_IMAGE_URL, savedMember);
+        maps.save(luther);
+
+        Setting beSetting = Setting.builder()
+                .availableStartTime(BE_AVAILABLE_START_TIME)
+                .availableEndTime(BE_AVAILABLE_END_TIME)
+                .reservationTimeUnit(BE_RESERVATION_TIME_UNIT)
+                .reservationMinimumTimeUnit(BE_RESERVATION_MINIMUM_TIME_UNIT)
+                .reservationMaximumTimeUnit(BE_RESERVATION_MAXIMUM_TIME_UNIT)
+                .reservationEnable(BE_RESERVATION_ENABLE)
+                .enabledDayOfWeek(BE_ENABLED_DAY_OF_WEEK)
+                .build();
+
+        Space be = Space.builder()
+                .name(BE_NAME)
+                .color(BE_COLOR)
+                .description(BE_DESCRIPTION)
+                .area(SPACE_DRAWING)
+                .setting(beSetting)
+                .map(luther)
+                .build();
+
+        spaces.save(be);
+
+        if (isReservationExists) {
+            Reservation beAmZeroOne = Reservation.builder()
+                    .startTime(BE_AM_TEN_ELEVEN_START_TIME)
+                    .endTime(BE_AM_TEN_ELEVEN_END_TIME)
+                    .description(BE_AM_TEN_ELEVEN_DESCRIPTION)
+                    .userName(BE_AM_TEN_ELEVEN_USERNAME)
+                    .password(BE_AM_TEN_ELEVEN_PW)
+                    .space(be)
+                    .build();
+
+            reservations.save(beAmZeroOne);
+        }
+
+        // when
+        Boolean hasAnyReservations = reservations.existsReservationsByMember(savedMember);
+
+        // then
+        AssertionsForClassTypes.assertThat(hasAnyReservations).isEqualTo(isReservationExists);
     }
 }
