+++ backend/src/test/java/com/woowacourse/zzimkkong/service/MemberServiceTest.java
@@ -85,7 +85,7 @@ class MemberServiceTest extends ServiceTest {
     void deleteMember() {
         // given
         Member member = new Member(1L, EMAIL, PW, ORGANIZATION);
-        given(members.existsReservationsByMember(any(Member.class)))
+        given(reservations.existsReservationsByMember(any(Member.class)))
                 .willReturn(false);
 
         // when, then
@@ -97,7 +97,7 @@ class MemberServiceTest extends ServiceTest {
     void deleteMemberFailWhenAnyReservationsExists() {
         // given
         Member member = new Member(1L, EMAIL, PW, ORGANIZATION);
-        given(members.existsReservationsByMember(any(Member.class)))
+        given(reservations.existsReservationsByMember(any(Member.class)))
                 .willReturn(true);
 
         // when, then
